:white_check_mark: tests(P.O): centralizing the configurations
Mapeando o arquivo Base para centralizar a configuração do Selenide e a implementação do Page Object, facilitando assim a manutenção futura.
